Implementado simulação de data e aumentado o progresso no metodo de registrar um emprestimo

+++ src/br/usp/icmc/ssc0103/Community.java
@@ -1,7 +1,7 @@
 package br.usp.icmc.ssc0103;
 
 public class Community extends User {
-    Community(String name, int remainingQuota, int time, int code) {
+    Community(String name, int remainingQuota, long time, int code) {
         this.name = name;
         this.remainingQuota = remainingQuota;
         this.time = time;
