+++ src/br/usp/icmc/ssc0103/Professor.java
@@ -1,7 +1,7 @@
 package br.usp.icmc.ssc0103;
 
 public class Professor extends User {
-    Professor(String name, int remainingQuota, int time, int code) {
+    Professor(String name, int remainingQuota, long time, int code) {
         this.name = name;
         this.remainingQuota = remainingQuota;
         this.time = time;
