+++ src/br/usp/icmc/ssc0103/Student.java
@@ -1,7 +1,7 @@
 package br.usp.icmc.ssc0103;
 
 public class Student extends User {
-    Student(String name, int remainingQuota, int time, int code) {
+    Student(String name, int remainingQuota, long time, int code) {
         this.name = name;
         this.remainingQuota = remainingQuota;
         this.time = time;
